requirements.txt
fastapi[standard]
uvicorn
SQLAlchemy
pydantic
Jinja2
python-multipart
Pillow
langchain
langchain-community
langchain-google-genai
pydantic
tavily-python
rapidfuzz
websockets
google-genai
openai-agents